Add repository infrastructure and CI workflows
Set up proper repository structure following rs-soroban-sdk patterns:

Build & CI:
- Add rust-toolchain.toml
- Add Makefile with build, test, fmt, check, doc targets
- Add GitHub Actions CI workflow (rust.yml)
- Add publish.yml for crates.io release automation

Documentation:
- Add Apache-2.0 LICENSE
- Add CONTRIBUTING.md with developer setup guide
- Add RELEASING.md with release process
- Add .github/docs/makefile-reference.md

Code fixes:
- Remove unnecessary `as u32` casts in sponge.rs (clippy)
- Update soroban-sdk from git dep to "=25.0.0-rc.2"

The crate is now ready for crates.io publishing.

+++ src/poseidon/sponge.rs
@@ -222,7 +222,7 @@ where
         for i in 0..inputs.len() {
             let v = inputs.get_unchecked(i);
             assert!(v < modulus, "input exceeds field modulus");
-            self.state.set(i as u32 + CAPACITY, v);
+            self.state.set(i + CAPACITY, v);
         }
     }
